# Increase laziness of builtin access in `fastReduce`