@gallais 2023-04-30
> There's probably work which could be done here to index builtins with an enum rather than string constants (which would help a little with lookup times), but that's a much more invasive change, so feel it's probably worth me rasing and discussing that first before I jump in :).

It would also hopefully help with maintenance by getting the coverage checker to shout at us whenever we add a builtin and forget to extend the appropriate functions. So :+1: from me.